// src/app/models/branch.ts
export interface Branch {
  id: number;
  name: string;
  description: string;
  countryId: number;
  cityId: number;
  isAirport: boolean;
  isActive: boolean;
  isPleaseOfPurchase: boolean;
  isBranch: boolean;
  areaId: number;
  address: string;
  imagePath: string;
  code: string;
  email: string;
  phone: string;
  phone2: string;
  coordination: string;
  order: number;
}